requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.30
python-dotenv>=1.0.0
requests>=2.31.0
typing-extensions>=4.5.0
langchain-google-genai>=1.0.0 
google-generativeai>=0.3.0     
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6